requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
boto3==1.34.127
botocore==1.34.127
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
Django==5.0.4
django-cleanup==8.1.0
django-crum==0.7.9
django-environ==0.11.2
django-extensions==3.2.3
django-storages==1.14.3
django-widget-tweaks==1.5.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.13
langchain-text-splitters==0.2.2
langsmith==0.1.83
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.35.7
orjson==3.10.5
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.1
pillow==10.4.0
pinecone-client==4.1.1
pinecone-plugin-interface==0.0.7
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.0
tenacity==8.4.2
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.1
yarl==1.9.4
